Use "Totals" for "Site totals" rather than "totals for every path"
A "unique visitor" is always "an unique visitor to this path". That's
fine behaviour for the paths overview, but not really all that useful
for the "Totals" chart. The way this works right now is that it counts
the number of unique visitors per path (so visiting /a and /b would show
up as *2*).

This changes that to make the Totals chart count only the total visitors
to your *site*. I don't think the current behaviour is ever really
useful. I initially added an option to toggle the behaviour, but then I
figured that was just unnecessary.

Fixes #471
Fixes #542